ci: :zap: Slider die casting transfer edit/add/delete  & order slider update
