+++ src/pages/Slider/DieCasting/Transfer/AgainstStock/index.jsx
@@ -1,0 +1,210 @@
+import { Suspense } from '@/components/Feedback';
+import ReactTable from '@/components/Table';
+import { useAccess } from '@/hooks';
+import {
+	useSliderDieCastingStock,
+	useSliderDieCastingTransferAgainstStock,
+} from '@/state/Slider';
+import { DateTime, EditDelete } from '@/ui';
+import PageInfo from '@/util/PageInfo';
+import { lazy, useEffect, useMemo, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+const AddOrUpdate = lazy(() => import('./AddOrUpdate'));
+
+const DeleteModal = lazy(() => import('@/components/Modal/Delete'));
+
+const Index = () => {
+	const navigate = useNavigate();
+	const { data, isLoading, url, deleteData } =
+		useSliderDieCastingTransferAgainstStock();
+	const { invalidateQuery } = useSliderDieCastingStock();
+	const info = new PageInfo(
+		'Transfer_by_stock',
+		url,
+		'slider__die_casting_transfer_by_stock'
+	);
+	const haveAccess = useAccess('slider__die_casting_transfer');
+
+	const columns = useMemo(
+		() => [
+			{
+				accessorKey: 'name',
+				header: 'Name',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'die_casting_uuid',
+				header: 'die_casting_uuid',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'quantity',
+				header: 'quantity',
+				enableColumnFilter: false,
+				cell: (info) => Number(info.getValue()),
+			},
+			{
+				accessorKey: 'end_type_name',
+				header: 'End Type',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'item_name',
+				header: 'Item',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'puller_link_name',
+				header: 'Puller Link',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'puller_type_name',
+				header: 'Puller Type',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'zipper_number_name',
+				header: 'Zipper No.',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'slider_body_shape_name',
+				header: 'Slider Body',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'created_by_name',
+				header: 'Created By',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'created_at',
+				header: 'Created At',
+				enableColumnFilter: false,
+				filterFn: 'isWithinRange',
+				cell: (info) => {
+					return <DateTime date={info.getValue()} />;
+				},
+			},
+			{
+				accessorKey: 'updated_at',
+				header: 'Updated',
+				enableColumnFilter: false,
+				cell: (info) => {
+					return <DateTime date={info.getValue()} />;
+				},
+			},
+			{
+				accessorKey: 'remarks',
+				header: 'Remarks',
+				enableColumnFilter: false,
+				cell: (info) => info.getValue(),
+			},
+			{
+				accessorKey: 'action',
+				header: 'Action',
+				enableColumnFilter: false,
+				hidden: !haveAccess.includes('update'),
+				cell: (info) => {
+					return (
+						<EditDelete
+							idx={info.row.index}
+							handelUpdate={handelUpdate}
+							handelDelete={() => handelDelete(info.row.id)}
+							showDelete={haveAccess.includes('delete')}
+							showUpdate={haveAccess.includes('update')}
+						/>
+					);
+				},
+			},
+		],
+		[data]
+	);
+
+	// Fetching data from server
+	useEffect(() => {
+		document.title = info.getTabName();
+	}, []);
+
+	// Add
+	const handelAdd = () => navigate('/slider/die-casting/transfer/entry');
+
+	// Update
+	const [update, setUpdate] = useState({
+		uuid: null,
+	});
+
+	const handelUpdate = (idx) => {
+		setUpdate((prev) => ({
+			...prev,
+			uuid: data[idx]?.uuid,
+		}));
+
+		window[info.getAddOrUpdateModalId()].showModal();
+	};
+
+	// Delete
+	const [deleteItem, setDeleteItem] = useState({
+		itemId: null,
+		itemName: null,
+	});
+	const handelDelete = (idx) => {
+		setDeleteItem((prev) => ({
+			...prev,
+			itemId: data[idx].uuid,
+			itemName: data[idx].name,
+		}));
+
+		window[info.getDeleteModalId()].showModal();
+	};
+
+	if (isLoading)
+		return <span className='loading loading-dots loading-lg z-50' />;
+
+	return (
+		<>
+			<ReactTable
+				title='Die Casting > Stock'
+				handelAdd={handelAdd}
+				accessor={haveAccess.includes('create')}
+				data={data}
+				columns={columns}
+			/>
+
+			<Suspense>
+				<AddOrUpdate
+					modalId={info.getAddOrUpdateModalId()}
+					{...{
+						update,
+						setUpdate,
+					}}
+				/>
+			</Suspense>
+			<Suspense>
+				<DeleteModal
+					modalId={info.getDeleteModalId()}
+					title={info.getTitle()}
+					{...{
+						deleteItem,
+						setDeleteItem,
+						url,
+						deleteData,
+						invalidateQuery,
+					}}
+				/>
+			</Suspense>
+		</>
+	);
+};
+
+export default Index;
